app/requirements.txt
eclipse-zenoh=0.11.0
Flask
Flask-SocketIO